requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pydantic==2.11.7
groq==1.1.1
python-dotenv==1.1.1
chromadb==1.5.8
redis==7.4.0
pytest==9.0.3
requests==2.32.3
numpy==2.2.3
